mlapiservertutorial/requirements.txt
alabaster==0.7.10
appnope==0.1.0
astroid==1.5.3
Babel==2.5.0
bleach==1.5.0
boto3==1.4.5
botocore==1.5.92
certifi==2016.2.28
chardet==3.0.4
colorlover==0.2.1
cufflinks==0.12.0
cycler==0.10.0
decorator==4.1.2
Django==1.11.5
djangorestframework==3.6.4
docutils==0.14
entrypoints==0.2.3
html5lib==0.999
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.0
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.4.0
notebook==5.0.0
numpy==1.13.1
numpydoc==0.7.0
pandas==0.20.3
pandocfilters==1.4.2
patsy==0.4.1
pexpect==4.2.1
pickleshare==0.7.4
plotly==2.0.15
prompt-toolkit==1.0.15
psutil==5.2.2
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.14.2
rope==0.10.7
rope-py3k==0.9.4.post1
s3transfer==0.1.10
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.3
statsmodels==0.8.0
terminado==0.6
testpath==0.3
tornado==4.5.2
traitlets==4.3.2
wcwidth==0.1.7
wrapt==1.10.11
